refactor(brain): add JarvisBrainProviders and move the OpenAI adapter out of Core
Extract the concrete OpenAI Responses adapter (OpenAIBrainClient) and its
HTTP permanence classification (BrainFailure+OpenAI) into the new
JarvisBrainProviders library target, which depends inward on JarvisCore.
Core keeps the provider-neutral brain domain — BrainClient/BrainConversation,
BrainTarget/BrainRoute, BrainFailure, model catalog, workload timeouts, and
tool-invocation parsing — and JarvisApp composes providers at Start.

One Core symbol widens to public: BrainFailure.init(_:), the provider
adapters' unknown-failure-to-temporary classification entry point. The
coaching parity harness keeps composing the kernel with the real OpenAI
adapter, so JarvisCoreTests links the provider target for that harness.
Route and attempt semantics are untouched; the harness's route-transition
snapshot pins them. The local-agent CLI subtree follows in #206.

Records the Phase 4 (OpenAI slice) scope contract in
wiki/lean-coaching-core.md per issue #205.

Closes #205
