+++ Tests/JarvisCoreTests/Brain/BrainFailureTests.swift
@@ -2,6 +2,10 @@ import Foundation
 import Testing
 @testable import JarvisCore
 
+/// The provider-neutral classification contract: everything a provider adapter has not proven
+/// anything about wraps as a recoverable `.temporary` missed turn. Each adapter's own permanence
+/// proof is tested beside that adapter (see `BrainFailureOpenAITests` in
+/// `JarvisBrainProvidersTests`).
 @Suite struct BrainFailureTests {
     @Test func unknownFutureErrorDefaultsToTemporary() {
         let failure = BrainFailure(NSError(
@@ -12,19 +16,14 @@ import Testing
         #expect(failure.detail == "new failure")
     }
 
-    @Test func transportAndServerFailuresShareTemporaryPolicy() {
+    @Test func transportFailuresShareTemporaryPolicy() {
         for error in [URLError(.timedOut), URLError(.networkConnectionLost)] {
             let failure = BrainFailure(error)
             #expect(failure.disposition == .temporary)
         }
-        for status in [408, 409, 500, 503, 599] {
-            let failure = BrainFailure.openAIHTTP(
-                status: status, errorCode: nil, errorType: nil, detail: "failed")
-            #expect(failure.disposition == .temporary)
-        }
     }
 
-    @Test func cliAndRateLimitFailuresRemainTemporary() {
+    @Test func cliFailuresRemainTemporary() {
         let errors: [Error] = [
             NSError(domain: AgentCLIProcessRunner.errorDomain, code: NSURLErrorTimedOut),
             NSError(domain: "CLIBrainClient", code: 1),
@@ -34,33 +33,9 @@ import Testing
             let failure = BrainFailure(error)
             #expect(failure.disposition == .temporary)
         }
-        let rateLimit = BrainFailure.openAIHTTP(
-            status: 429, errorCode: "rate_limit_exceeded", errorType: nil, detail: "limited")
-        #expect(rateLimit.disposition == .temporary)
     }
 
-    @Test func unknownAndRequestLocalHTTPFailuresRemainTemporary() {
-        for status in [400, 404, 418, 422, 423, 424, 425, 429] {
-            let failure = BrainFailure.openAIHTTP(
-                status: status, errorCode: "future_code", errorType: "future_type", detail: "failed")
-            #expect(failure.disposition == .temporary)
-        }
-    }
-
-    @Test func onlyProvenOrExplicitPermanentFailuresArePermanent() {
-        for status in [401, 402, 403] {
-            #expect(BrainFailure.openAIHTTP(
-                status: status, errorCode: nil, errorType: nil, detail: "failed"
-            ).disposition == .permanent)
-        }
-        for code in ["invalid_api_key", "insufficient_quota", "model_not_found"] {
-            #expect(BrainFailure.openAIHTTP(
-                status: 429, errorCode: code, errorType: nil, detail: "failed"
-            ).disposition == .permanent)
-        }
-        #expect(BrainFailure.openAIHTTP(
-            status: 429, errorCode: nil, errorType: "authentication_error", detail: "failed"
-        ).disposition == .permanent)
+    @Test func explicitPermanentFailureSurvivesRewrapping() {
         let explicit = BrainFailure(
             disposition: .permanent, detail: "provider proved authentication is invalid")
         #expect(BrainFailure(explicit) == explicit)
